+++ src/config/mod.rs
@@ -1,10 +1,9 @@
 mod date_format;
 mod model;
-mod project;
 mod search;
 mod timezone;
 
 // Re-export public items
+pub use crate::post::find_project_root;
 pub use model::Config;
-pub use project::find_project_root;
 pub use search::{SearchConfig, SearchLanguageConfig};